pyproject.toml
[project]
name = "privilix"
version = "1.0.01"
description = "The only moderation tool you need for discord"
authors = [
    {name = "ayushdotdev",email = "ayush438500@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "babel (>=2.17.0,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "aerich (>=0.9.2,<0.10.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.services.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
